Style the connector
