+++ frontend/src/metabase/admin/datamodel/metadata/components/FieldGeneralSettings/FieldGeneralSettings.tsx
@@ -201,7 +201,6 @@ const FieldTypeSection = ({
         onUpdateField={(_field, updates) => {
           onUpdateField(field, updates);
         }}
-        hasSeparator
       />
     </MetadataSection>
   );
